and its working
managed to get a PostgreSQL server running, which allowed me to actually find the bug and fix it. (it was a lowercase A, at CreatedAt in the TaskRepository, in case you're wondering)
Didn't manually test it yet, but I'm getting a failed test at the Update endpoint. Not sure what's up with that right now and would appreciate feedback